+++ binobj/typedefs.py
@@ -3,24 +3,20 @@
 from __future__ import annotations
 
 import typing
+from collections.abc import Callable
 from collections.abc import Mapping
 from collections.abc import MutableMapping
 from typing import Any
-from typing import Callable
 from typing import Literal
-from typing import Optional
-from typing import TypeVar
 from typing import Union
 
 
 if typing.TYPE_CHECKING:  # pragma: no cover
-    from typing_extensions import TypeAlias
+    from typing import TypeAlias
 
     from binobj.fields.base import Field
     from binobj.structures import Struct
 
-
-T_co = TypeVar("T_co", covariant=True, bound=object)
 
 StrDict: TypeAlias = Mapping[str, Any]
 """Any mapping using strings as the keys."""
@@ -28,15 +24,13 @@ StrDict: TypeAlias = Mapping[str, Any]
 MutableStrDict: TypeAlias = MutableMapping[str, Any]
 """Any mutable mapping using strings for the keys."""
 
-FieldValidator: TypeAlias = Callable[["Field[Any]", Any], Optional[bool]]
+FieldValidator: TypeAlias = Callable[["Field[Any]", Any], bool | None]
 """A function that, given a field, returns True if it's valid and False otherwise."""
 
-MethodFieldValidator: TypeAlias = Callable[
-    ["Struct", "Field[Any]", Any], Optional[bool]
-]
+MethodFieldValidator: TypeAlias = Callable[["Struct", "Field[Any]", Any], bool | None]
 """A field validator that's a method in its containing struct."""
 
-StructValidator: TypeAlias = Callable[["Struct", StrDict], Optional[bool]]
+StructValidator: TypeAlias = Callable[["Struct", StrDict], bool | None]
 """A classmethod that validates an entire struct."""
 
 FieldOrName: TypeAlias = Union[str, "Field[Any]"]